Added unit tests for verifying stack tracking
Currently, there is no support for polymorphic operations.
And, JavaCoreCodeWriterImpl needs to have the order of its unstack calls in the various methods reversed.

Moved Operation into the net.dougqh.jak.operations package.
Moved Types into net.dougqh.jak.types package.

Fixed a defect with the stack tracking in JavaCoreCodeWriterImpl.laload.
Fixed incorrect annotations on methods in JavaCoreCodeWriter.
Fixed improper implementations of getStackResultTypes() in Iaload classes.

+++ net.dougqh.jak/src/net/dougqh/jak/operations/Istore.java
@@ -1,7 +1,6 @@
 package net.dougqh.jak.operations;
 
 import net.dougqh.jak.JavaCoreCodeWriter;
-import net.dougqh.jak.Operation;
 
 public final class Istore extends Operation {
 	public static final String ID = "istore";
@@ -34,7 +33,7 @@ public final class Istore extends Operation {
 	
 	@Override
 	public final Class< ? >[] getStackOperandTypes() {
-		return new Class< ? >[] { long.class };
+		return new Class< ? >[] { int.class };
 	}
 	
 	@Override
